Additonal float conversions

+++ CHC/CHC_Base/CHC_Editor.cpp
@@ -743,13 +743,13 @@ void CHC_Editor::audioSettings()
 			for (size_t index = 0; index < 8; index++)
 			{
 				printf("%s     %zu    ||", g_global.tabs.c_str(), index + 1);
-				printf("%7lu  ||%7g%%   ||", m_song->m_audio[index].volume, ((float)m_song->m_audio[index].volume) * 100 / 32767);
+				printf("%7lu  ||%7g%%   ||", m_song->m_audio[index].volume, m_song->m_audio[index].volume * 100.0f / 32767);
 				switch (m_song->m_audio[index].pan)
 				{
 				case 0:     printf("   Left   |           ||\n"); break;
 				case 16383: printf("        Center        ||\n"); break;
 				case 32767: printf("          |   Right   ||\n"); break;
-				default: printf("%8g%% |%8g%% ||\n", 100 - (((float)m_song->m_audio[index].pan) * 100 / 32767), ((float)m_song->m_audio[index].pan) * 100 / 32767);
+				default: printf("%8g%% |%8g%% ||\n", 100 - (m_song->m_audio[index].pan * 100.0f / 32767), m_song->m_audio[index].pan * 100.0f / 32767);
 				}
 			}
 			printf("%s==========================================================||\n", g_global.tabs.c_str());
@@ -766,7 +766,7 @@ void CHC_Editor::audioSettings()
 			for (size_t index = 2; index < 8; index++)
 			{
 				printf("%s     %zu    ||", g_global.tabs.c_str(), index + 1);
-				printf("%7lu  ||%7g%%   ||", m_song->m_audio[index].volume, ((float)m_song->m_audio[index].volume) * 100 / 32767);
+				printf("%7lu  ||%7g%%   ||", m_song->m_audio[index].volume, m_song->m_audio[index].volume * 100.0f / 32767);
 			}
 			printf("%s==================================||\n", g_global.tabs.c_str());
 			printf("%sChoose the channel you wish to edit [Type 'Q' to exit audio settings]\n", g_global.tabs.c_str());
@@ -849,7 +849,7 @@ void CHC_Editor::audioSettings()
 							case 0: printf("Left (0)\n"); break;
 							case 16383: printf("Center (16383)\n"); break;
 							case 32767: printf("Right (32767)\n"); break;
-							default: printf("%g%% Left | %g%% Right (%lu)\n", 100 - (m_song->m_audio[channel].pan * 100.0 / 32767), m_song->m_audio[channel].pan * 100.0 / 32767, m_song->m_audio[channel].pan);
+							default: printf("%g%% Left | %g%% Right (%lu)\n", 100 - (m_song->m_audio[channel].pan * 100.0f / 32767), m_song->m_audio[channel].pan * 100.0f / 32767, m_song->m_audio[channel].pan);
 							}
 							printf("%sInput: ", g_global.tabs.c_str());
 							switch (GlobalFunctions::valueInsert(m_song->m_audio[channel].pan, false, 0UL, 32767UL, "b"))
@@ -1851,7 +1851,7 @@ bool CHC_Editor::reorganize(SongSection& section)
 					if (ntIndex == startingIndex)
 					{
 						if (notes[pl][startingIndex].first >= 2 * SAMPLES_PER_BEAT)
-							currentChart->setPivotTime((long)round(notes[pl][startingIndex].first - SAMPLES_PER_BEAT));
+							currentChart->setPivotTime((long)roundf(notes[pl][startingIndex].first - SAMPLES_PER_BEAT));
 						else
 							currentChart->setPivotTime(notes[pl][startingIndex].first >> 1);
 						notes[pl][startingIndex].second->setPivotAlpha(notes[pl][startingIndex].first - currentChart->m_pivotTime);
